maj
centralisation du security config
erreur lecture Risk diabete
erreur endpoint frontend

+++ doctor-notes/src/main/java/com/sprint2/doctor_notes/controller/NoteController.java
@@ -41,7 +41,7 @@ public class NoteController {
     }
 
     @GetMapping("/{patientId}")
-    public List<Note> getNotes(@PathVariable Long patientId) {
+    public List<Note> getNotes(@PathVariable int patientId) {
         return noteService.getNotesByPatient(patientId);
     }
 }
